Refactor ftl.run
