+++ test_runner/src/main/kotlin/ftl/args/ShardChunks.kt
@@ -1,0 +1,3 @@
+package ftl.args
+
+typealias ShardChunks = List<List<String>>
