--- test_runner/src/main/kotlin/ftl/run/GenericTestRunner.kt
@@ -1,91 +1,0 @@
-package ftl.run
-
-import com.google.api.services.testing.model.TestMatrix
-import ftl.args.IArgs
-import ftl.config.FtlConstants
-import ftl.config.FtlConstants.useMock
-import ftl.json.MatrixMap
-import ftl.json.SavedMatrix
-import ftl.util.StopWatch
-import ftl.util.uniqueObjectName
-import java.io.File
-
-object GenericTestRunner {
-    fun beforeRunTests(args: IArgs): Pair<StopWatch, String> {
-        println("RunTests")
-        val stopwatch = StopWatch().start()
-        TestRunner.assertMockUrl()
-
-        val runGcsPath = args.resultsDir ?: uniqueObjectName()
-
-        // Avoid spamming the results/ dir with temporary files from running the test suite.
-        if (useMock) {
-            Runtime.getRuntime().addShutdownHook(Thread {
-                File(args.localResultDir, runGcsPath).deleteRecursively()
-            })
-        }
-
-        if (args.useLocalResultDir()) {
-            // Only one result is stored when using --local-result-dir
-            // Delete any old results if they exist before storing new ones.
-            File(args.localResultDir).deleteRecursively()
-        }
-
-        return stopwatch to runGcsPath
-    }
-
-    fun afterRunTests(
-        jobs: List<TestMatrix>,
-        runGcsPath: String,
-        stopwatch: StopWatch,
-        config: IArgs
-    ): MatrixMap {
-        val savedMatrices = mutableMapOf<String, SavedMatrix>()
-
-        jobs.forEach { matrix ->
-            val matrixId = matrix.testMatrixId
-            savedMatrices[matrixId] = SavedMatrix(matrix)
-        }
-
-        val matrixMap = MatrixMap(savedMatrices, runGcsPath)
-        TestRunner.updateMatrixFile(matrixMap, config)
-        TestRunner.saveConfigFile(matrixMap, config)
-
-        println(FtlConstants.indent + "${savedMatrices.size} matrix ids created in ${stopwatch.check()}")
-        val gcsBucket = "https://console.developers.google.com/storage/browser/" +
-                config.resultsBucket + "/" + matrixMap.runPath
-        println(FtlConstants.indent + gcsBucket)
-        println()
-
-        return matrixMap
-    }
-
-    private fun s(amount: Int): String {
-        return if (amount > 1) {
-            "s"
-        } else {
-            ""
-        }
-    }
-
-    fun beforeRunMessage(args: IArgs, testShardChunks: List<List<String>>): String {
-        val runCount = args.repeatTests
-        val shardCount = testShardChunks.size
-        val testsCount = testShardChunks.sumBy { it.size }
-
-        val result = StringBuilder()
-        result.appendln(
-            "  $testsCount test${s(testsCount)} / $shardCount shard${s(shardCount)}"
-        )
-
-        if (runCount > 1) {
-            result.appendln("  Running ${runCount}x")
-            val runDevices = runCount * shardCount
-            val runTests = runCount * testsCount
-            result.appendln("    $runDevices total shard${s(runDevices)}")
-            result.appendln("    $runTests total test${s(runTests)}")
-        }
-
-        return result.toString()
-    }
-}
