+++ test_runner/src/main/kotlin/ftl/run/common/PrettyPrint.kt
@@ -1,0 +1,6 @@
+package ftl.run.common
+
+import com.google.gson.Gson
+import com.google.gson.GsonBuilder
+
+val prettyPrint: Gson = GsonBuilder().setPrettyPrinting().create()!!
