+++ test_runner/src/main/kotlin/ftl/run/model/TestResult.kt
@@ -1,0 +1,6 @@
+package ftl.run.model
+
+import ftl.json.MatrixMap
+import ftl.args.ShardChunks
+
+internal typealias TestResult = Pair<MatrixMap, ShardChunks>
