+++ test_runner/src/test/kotlin/Debug.kt
@@ -10,17 +10,19 @@ fun main() {
     // run "gradle check" to generate required fixtures
     val projectId = System.getenv("FLANK_PROJECT_ID")
         ?: "YOUR PROJECT ID"
-    val quantity = "single"
-    val type = "success"
+    val quantity = "multiple"
+    val type = "flaky"
 
     // Bugsnag keeps the process alive so we must call exitProcess
     // https://github.com/bugsnag/bugsnag-java/issues/151
-    jvmHangingSafe { CommandLine(Main()).execute(
-        "--debug",
-        "firebase", "test",
-        "android", "run",
-        "--dry",
-        "-c=src/test/kotlin/ftl/fixtures/test_app_cases/flank-$quantity-$type.yml",
-        "--project=$projectId"
-    ) }
+    jvmHangingSafe {
+        CommandLine(Main()).execute(
+//        "--debug",
+            "firebase", "test",
+            "android", "run",
+//        "--dry",
+            "-c=src/test/kotlin/ftl/fixtures/test_app_cases/flank-$quantity-$type.yml",
+            "--project=$projectId"
+        )
+    }
 }
